refactor(session): widen scroll offset from u16 to u32
Support sessions with 10M+ token threads by upgrading the scroll
offset type from u16 to u32 across event handling, rendering,
scrollbar, and anchor-jump paths. Also add a fallback in
build_lines_viewport to rebuild all lines when line_offsets is
stale, preventing a blank viewport after cache invalidation.

+++ src/function/mod.rs
@@ -1893,7 +1893,7 @@ impl App {
     /// not coast. The integer offset is written to `session.scroll` so
     /// the existing render path picks it up, and the render cache is
     /// invalidated so the change is visible on the next frame.
-    pub fn set_scroll_anchored(&mut self, value: u16) {
+    pub fn set_scroll_anchored(&mut self, value: u32) {
         self.session_scroll.snap(value as f32);
         self.session.scroll = value;
     }
